Spells are now also checked for wall passability in indoor maps

+++ src/stratagus/player.cpp
@@ -2227,14 +2227,14 @@ std::string CPlayer::HasFailedQuest(CQuest *quest) // returns the reason for fai
 		if (std::get<0>(this->QuestBuildUnits[i]) == quest && std::get<2>(this->QuestBuildUnits[i]) > 0) {
 			bool has_builder = false;
 			CUnitType *type = std::get<1>(this->QuestBuildUnits[i]);
-			if (type->BoolFlag[BUILDING_INDEX].value && type->Slot < AiHelpers.Build.size()) {
+			if (type->BoolFlag[BUILDING_INDEX].value && type->Slot < (int) AiHelpers.Build.size()) {
 				for (size_t j = 0; j < AiHelpers.Build[type->Slot].size(); ++j) {
 					if (this->UnitTypesCount[AiHelpers.Build[type->Slot][j]->Slot] > 0) {
 						has_builder = true;
 						break;
 					}
 				}
-			} else if (!type->BoolFlag[BUILDING_INDEX].value && type->Slot < AiHelpers.Train.size()) {
+			} else if (!type->BoolFlag[BUILDING_INDEX].value && type->Slot < (int) AiHelpers.Train.size()) {
 				for (size_t j = 0; j < AiHelpers.Train[type->Slot].size(); ++j) {
 					if (this->UnitTypesCount[AiHelpers.Train[type->Slot][j]->Slot] > 0) {
 						has_builder = true;
@@ -2258,14 +2258,14 @@ std::string CPlayer::HasFailedQuest(CQuest *quest) // returns the reason for fai
 			}
 			
 			CUnitType *type = UnitTypes[unit_type_id];
-			if (type->BoolFlag[BUILDING_INDEX].value && type->Slot < AiHelpers.Build.size()) {
+			if (type->BoolFlag[BUILDING_INDEX].value && type->Slot < (int) AiHelpers.Build.size()) {
 				for (size_t j = 0; j < AiHelpers.Build[type->Slot].size(); ++j) {
 					if (this->UnitTypesCount[AiHelpers.Build[type->Slot][j]->Slot] > 0) {
 						has_builder = true;
 						break;
 					}
 				}
-			} else if (!type->BoolFlag[BUILDING_INDEX].value && type->Slot < AiHelpers.Train.size()) {
+			} else if (!type->BoolFlag[BUILDING_INDEX].value && type->Slot < (int) AiHelpers.Train.size()) {
 				for (size_t j = 0; j < AiHelpers.Train[type->Slot].size(); ++j) {
 					if (this->UnitTypesCount[AiHelpers.Train[type->Slot][j]->Slot] > 0) {
 						has_builder = true;
